0.20.0.4: Move class Citizen to Citizen.cpp/h
